Clean up sloppy code.
How did this code ever make it past code review, in the first place?
